+++ src/style.css
@@ -69,7 +69,7 @@ html {
   border-bottom: #e6e4e4 solid 0.1rem;
 }
 
-.el-1 {
+.fas {
   margin-left: 1rem;
   margin-right: 1rem;
   color: #cecdcd;
